Bug fix
When a failureCallable is used, there's no CSRF tokens set when
something is rendered through there. This is caused because the
validateToken removes any set token, so it needs to be generated before
the $failureCallable returns. Moved the code to generate and assign a
new token to a separate method for this.