Limit storage array size
Remove the oldest tokens from the storage array so that there are never
more than storageLimit tokens in the array. This is required as a token
is generated every request and so most will never be used.

I've picked 200 as an arbitrary number on the assumption that for any
given form, the user is unlikely to make 200 more requests before
submitting it.