Improve test coverage for \Slim\Csrf\Guard::enforceStorageLimit()
During a refacotiring of this method, the early-return `if` block was not tested properly.
There is a new test \Slim\Tests\Csrf\GuardTest::testNotEnforceStorageLimitWithArrayWhenLimitIsZero that makes sure the storage array provided remains the same after a token is processed when the storage limit is set to zero, which should trigger the early-return `if` block, this adding more coverage.